pyproject.toml
[tool.poetry]
name = "crewai-stock-analyzer-aims"
version = "0.1.0"
description = "A stock analyzer built with CrewAI for an AI Makerspace Demo"
authors = ["Chris Alexiuk <chris@aimakerspace.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.32.2"
llama-parse = "^0.4.4"
llama-index-core = "^0.10.50.post1"
llama-index-readers-file = "^0.1.25"
python-dotenv = "^1.0.1"
langchain = "0.1.20"
google-search-results = "^2.4.2"
langchain-community = "0.0.38"
anthropic = "^0.29.0"
langchain-anthropic = "0.1.10"
yfinance = "^0.2.40"
sec-api = "^1.0.18"
unstructured = "^0.14.8"
langchain-openai = "0.1.7"
faiss-cpu = "^1.8.0.post1"
setuptools = "^75.8.0"
plotly = "^6.0.0"
notebook = "^7.3.2"
binance = "^0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

